backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Handling
python-multipart==0.0.9
boto3==1.34.50

# Task Queue
celery==5.3.6
redis==5.0.2

# AI/ML
openai-whisper
faster-whisper==1.0.3
groq==0.5.0
spacy==3.7.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-
google-api-python-client==2.118.0

# Utils
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Dev
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
